# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Mpango ERP API
  description: |
    Multi-tenant ERP system for African wholesale-retail operations.

    ## Authentication
    All endpoints except `/auth/login` require Bearer token authentication.
    The JWT token contains `tenant_id`, `tenant_schema`, and `user_id` claims.

    ## Multi-Tenancy
    - Login requires `tenant_code` to identify the wholesaler
    - All subsequent requests use tenant info from JWT claims
    - Data is isolated per tenant schema

    ## Authorization
    Endpoints are protected by RBAC permissions as defined in RBAC Matrix (MVP).
  version: 1.0.0
  contact:
    name: Mpango ERP Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.mpango.io/api/v1
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Authenticate user and get tokens
      description: |
        Multi-tenant login flow:
        1. Validate tenant_code against public.wholesalers
        2. Derive tenant_schema from wholesaler.id
        3. Authenticate user in tenant schema
        4. Return JWT with tenant claims
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request or inactive user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      summary: Get current authenticated user info
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # USER ENDPOINTS
  # ============================================================
  /users:
    get:
      tags: [users]
      operationId: listUsers
      summary: List users with pagination
      description: Requires `users:read` permission
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [users]
      operationId: createUser
      summary: Create a new user
      description: Requires `users:create` permission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{user_id}:
    get:
      tags: [users]
      operationId: getUser
      summary: Get user by ID
      description: Requires `users:read` permission
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [users]
      operationId: updateUser
      summary: Update user
      description: Requires `users:update` permission
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [users]
      operationId: deleteUser
      summary: Soft delete user (deactivate)
      description: Requires `users:deactivate` permission
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '204':
          description: User deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing claims:
        - user_id: UUID
        - tenant_id: UUID
        - tenant_schema: string (e.g., t_<uuid>)
        - exp: expiration timestamp

  # ----------------------------------------------------------
  # Parameters
  # ----------------------------------------------------------
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    UserIdParam:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ----------------------------------------------------------
  # Schemas
  # ----------------------------------------------------------
  schemas:
...
    Pagination:
      type: object
      required: [page, size, total, pages]
      properties:
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        pages:
          type: integer
          example: 10
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
        meta:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [success, error, timestamp]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Input validation failed
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: string
          format: date-time

    # ========== Payment ==========
...
    LoginRequest:
      type: object
      required: [tenant_code, email, password]
      properties:
        tenant_code:
          type: string
          description: Wholesaler code (e.g., ACME01)
          example: ACME01
          minLength: 1
          maxLength: 32
        email:
          type: string
          format: email
          example: admin@acme.com
        password:
          type: string
          format: password
          minLength: 8
          example: Admin123!
    LoginResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required: [access_token, refresh_token, token_type, user_id, tenant_id, tenant_schema]
          properties:
            access_token:
              type: string
              description: JWT access token
            refresh_token:
              type: string
              description: JWT refresh token
            token_type:
              type: string
              example: bearer
            user_id:
              type: string
              format: uuid
            tenant_id:
              type: string
              format: uuid
            tenant_schema:
              type: string
              example: t_11111111111111111111111111111111
        timestamp:
          type: string
          format: date-time
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    CurrentUserResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required: [id, email, tenant_id, tenant_schema, roles, permissions]
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            full_name:
              type: string
            tenant_id:
              type: string
              format: uuid
            tenant_schema:
              type: string
            roles:
              type: array
              items:
                type: string
              example: [admin]
            permissions:
              type: array
              items:
                type: string
              example: [users:read, users:create, orders:read]
        timestamp:
          type: string
          format: date-time

    # ========== User ==========
    UserCreateRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: newuser@acme.com
        password:
          type: string
          format: password
          minLength: 8
        full_name:
          type: string
          maxLength: 100
          example: John Doe
    UserUpdateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        full_name:
          type: string
          maxLength: 100
        is_active:
          type: boolean
    User:
      type: object
      required: [id, email, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        is_active:
          type: boolean
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/User'
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    UserListResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required: [items, pagination]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/User'
            pagination:
              $ref: '#/components/schemas/Pagination'
        timestamp:
          type: string
          format: date-time
...
    Role:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: admin
        description:
          type: string
...